update skip list
